Extract DiscountLineDetailsView to own file
